Refactored error handling.
Refactored the error handling.

If a C binding returns a non 0 return code duktape will generate an
appropriate Error() object and throw it so there is no need for us to
explicitly throw one.

Where functions are not C bindings, in a couple of cases, we now
explicitly create but do not throw an error object. The non 0 return
code is detected at the top level and the error reported.

Uncaught errors will always be handled at the top level. When JSE is
in CGI mode, this results in an HTTP status 500 error page being
returned identifying the exception.