Added basic HTTP error return function
Added a function to return a very basic HTTP error page. This is used
when jse is in CGI mode and an error hasn't already been handled. It
is necessary for fast CGI mode as the jse process does not exit on
error but continues to run so errors need to be returned as HTTP
statuses.