Added PosixError object and throw methods for it.
PosixError has been added. It can be constructed as follows:

  new PosixError(errno, message)

Where 'errno' is the POSIX error and 'message' is a descriptive message.
PosixErrors can be created and thrown in one step with the JS function
binding:

  throwPosixError(errno, message)

Where the parameters are as before. There is also an internal function
that does similarly:

  jse_throw_posix_error(ctx, errno, format, ...)

Where 'ctx' is the duktape context, 'errno' is the POSIX error, and format
is a printf style format string followed by parameters.

Macros have been added to throw common error types including PosixError
which report the message as a JSE_ERROR and throw the error in one go.