Added listDirectory() binding
The function listDirectory() returns an array of the files in the specified
directory.

If the directory does not exist a PosixError is thrown  with errno set to
ENOENT. A PosixError could be thrown due to other errors and errno should
be checked when one is caught.

If the directory specified is not actually a directory a URIError is
thrown.

Otherwise the files are returned as an array of strings in the order they are
read. Sorting and filtering can be done in JS afterwards.