Added CosaError object and throw methods for it.
CosaError has been added. It can be constructed as follows:

  new CosaError(errorCode, message)

Where 'errorCode' is the Cosa error and 'message' is a descriptive
message. CosaErrors can be created and thrown in one step with the
JS function binding:

  throwCosaError(errorCode, message)

Where the parameters are as before. There is also an internal function
that does similarly:

  jse_throw_cosa_error(ctx, error_code, format, ...)

Where 'ctx' is the duktape context, 'error_code' is the Cosa error, and
format is a printf style format string followed by parameters.

There is a macro which throws a CosaError and reports the message on debug
in one go.