errno does not need to be cached. Removing _errno copies.
Per the C standard errno is not reset by any functions. However it may be
overwritten by a subsequent error. Technically errno should be set to zero
before each function call that could change it however most functions
return a status that indicates an error occurred so we know when its been
set or not.