Implemented execProcess() to run external processes in JSE.
execProcess() runs an external process. The arguments comprise:

  string - the filename.
  array  - the argument array to the process. (optional)
  object - the environment object. (optional)

The filename must be a full path to the executable.

The argument array is an element for each argument to the process. The
first item should NOT be the path to the executable (per UNIX standards)
as this is done for you.

The environment object comprises an object with simple properties. The
environment is passed as an array of property=value pairs. For example:

{
   PATH = "/bin:/usr/bin",
   HOME = "/root"
}

will pass an array with "PATH=/bin:/usr/bin" as the first element and
"HOME=/root" as the second.

The process does NOT run in a shell. If you want a shell you have to
explicitly execute the shell.

The function returns an object with the following four properties:

  int    'pid'    - the PID of the process.
  int    'status' - the return status of the process.
  string 'stdout' - the standard out stream captured as a string.
  string 'stderr' - the standard error string captured as a string.

 Please enter the commit message for your changes. Lines starting