Fixed jse_read_fd() to return the total size read
jse_read_fd() was incorrectly returning the number of characters read
on the last read, not the total read (as it should).

I also noticed that functions that did not allocate memory blocks were
freeing them on error. It should be the responsibility of the caller
who original allocated the blocks on, to free them on error.