Added file upload support to JSE
File upload is supported when JSE is in HTTP request mode. When a form
has an <input type="file"> input the file is passed to JSE as a multipart
encoded request. The file is uploaded to the upload directory and an
associated file object is added to the Request.QueryParameters object.
The file object has the following properties:

  filename    - the filename
  contenttype - the file content type
  savepath    - the save path

The upload directory defaults to /var/jse/uploads but may be overridden
using the --upload-dir / -u command line option