jse_common.c: removed NUL character when reading from STDIN
When reading from stdin or a pipe there was an additional NUL character
being added. The intention was to have a 'safety' string terminator.
What actually happened was there was a NUL byte at the end of the input
which was sent to duktape and caused issues so this has been removed.

However this showed another bug. When the input is empty, the realloc()
actually freed the buffer generating a duplicate free() which is, of
course, very bad. So now the buffer is not realloc()ed if it no bytes
have been read.