jse_jscommon.c: Updated writeAsFile() to support buffer objects
The writeAsFile() binding now supports buffer objects. So if the object
being written is a duktape buffer, or a Uint8Array etc. it will be
written as raw data to the file. All other objects types behave as before.
I.e. by using their toString() method to convert to a string that is
written to the file.