jse_crypt.c: Added encryption and decryption binding functionality
The encrypt(), decryptToString() and decryptToBuffer() bindings now work.
They all take similar arguments.

 - The data to encrypt/decrypt
 - The key to use
 - The encryption/decryption method (optional)
 - The IV to use (optional)

When encrypting the data can be any value. Non buffer, non string values
are coerced to strings. Buffers are used as is.

When decryption the data is expected to be a buffer. Any other type is
an error.

If the encryption/decryption method is not specified then AES 256 ECB is
used. This does not require an IV.

The key and IV must be the right length for the encryption/decryption
method selected.

encrypt() returns a buffer object. decryptToString() returns a string
and decryptToBuffer() returns a buffer object.