Made jse_cosa_init() more robust and added try again feature
jse_cosa_init() tried to continue to initialise CCSP on NON RBUS builds
even when the initial CCSP bus initialisation failed. This may be the
cause of crashes. The code here has been improved.

Also if jse_cosa_init() fails (because we're trying to do it too early)
it will get tried again for each request. Once successful it is never
tried again. Failure to initialise on a request results in an HTTP status
500, Internal Server Error but JSE continues to run.