jse_crypt.c: added support for buffers as keys and ivs
Added support for buffers as keys and ivs. If the type of the key and/or
iv is 'buffer data' then the key and/or iv is assumed to be binary data
and the buffer length is used, otherwise it is assumed to be a string or
a type coorceable to a string in which case strlen() is used to calculate
its length.