jse_cosa.c: make local copies of strings from duk_safe_to_string()
If duk_safe_to_string() creates an internal string object when casting
or coercing a value to a string this can be short lived. This may be
a bug in Duktape. So use strdup() to make a copy of the string locally
freeing it when we are done with CCSP calls.