jse_main.c: added request body support to the Request object
HTTP requests that have mime types that are not supported by qdecoder now
have their request bodies returned as a property of the Request object.
QDecoder supports request mime types of "application/x-www-form-urlencoded"
or "multipart/form-data". For other types the request body is returned as
a property "body" in the Request object. The request body is only read for
HTTP methods of "PUT", "POST" and "PATCH".